Restructure project directories and files
Amendments to CMakeLists.txt, scripts/build.sh as required.

This is largely moving things around, especially making clear our public
header vs internal project headers;

also took the opportunity to
unify our source file includes under a configuration variable (QNN
delegate / GPU delegate files) to simplify the CMakeLists.txt.

Likewise, simplified platform targets with relevant lists of libraries,
modified conditionally, and single link command on properly configured
list.

Added requirements that QNN_DELEGATE, GPU_DELEGATE are always defined,
and have valid values of either 0 or 1.
