+++ CMakeLists.txt
@@ -4,7 +4,7 @@
 cmake_minimum_required(VERSION 3.22.1)
 
 # project name and language
-project(whisperax LANGUAGES CXX)
+project(whisperkit LANGUAGES CXX)
 
 if (NOT DEFINED TENSORFLOW_SOURCE_DIR)
     message(FATAL_ERROR  "TensorFlow path is not defined. Please use cmake -DTENSORFLOW_SOURCE_DIR=/path/to/tensorflow")
@@ -14,18 +14,35 @@ set(CMAKE_CXX_STANDARD 23)
 set(COMPILE_WARNING_AS_ERROR ON)
 set(CMAKE_BUILD_TYPE Release)
 
-set(AXIE_SRC_DIR ${CMAKE_SOURCE_DIR}/src)
-set(AXIE_INC_DIR ${CMAKE_SOURCE_DIR}/inc)
+set(WHISPERKIT_SRC_DIR ${CMAKE_SOURCE_DIR}/WhisperKit/src/)
 set(EXT_INC_DIR ${CMAKE_SOURCE_DIR}/external/inc)
-set(AXIE_CLI_DIR ${CMAKE_SOURCE_DIR}/cli)
+set(WHISPERKIT_CLI_DIR ${CMAKE_SOURCE_DIR}/cli)
+set(WHISPERKIT_PUBLIC_HEADERS ${CMAKE_SOURCE_DIR}/WhisperKit/)
+set(WHISPERKIT_EXTERNAL_HEADERS ${CMAKE_SOURCE_DIR}/WhisperKit/external/)
+
+file(GLOB_RECURSE WHISPERKIT_SOURCES ${WHISPERKIT_SRC_DIR}/*.cpp)
+
+set(WHISPERKIT_INTERNAL_HEADERS_DIRECTORIES
+  ${WHISPERKIT_SRC_DIR}/Core/
+  ${WHISPERKIT_SRC_DIR}/Models/
+  ${WHISPERKIT_SRC_DIR}/Text/
+  ${WHISPERKIT_SRC_DIR}/Audio/
+  ${WHISPERKIT_SRC_DIR}/
+)
+
+# QNN delegate legal only on Android with QCOM SoC;
+# GPU delegate legal on all Android;
+# linux configuration currently supports CPU only.
+OPTION(QNN_DELEGATE "QNN delegate" OFF) # Disabled by default
+OPTION(GPU_DELEGATE "GPU delegate" OFF) # Disabled by default
+SET(DELEGATE_ALLOWED_VALUES 0 1)
 
 if(${ANDROID})
   message(NOTICE  "Config build for Android ARM64..")
-  OPTION(QNN_DELEGATE "QNN delegate" OFF) # Disabled by default
   set(CMAKE_TOOLCHAIN_FILE ${ANDROID_NDK_ROOT}/build/cmake/android.toolchain.cmake)
   set(ANDROID_ABI arm64-v8a)
   set(EXT_LIB_DIR ${CMAKE_SOURCE_DIR}/external/libs/android)
-  find_library(android_log log)
+  find_library(ANDROID_LOG log)
 else()
   message(NOTICE  "Config build for linux..")
   set(EXT_LIB_DIR ${CMAKE_SOURCE_DIR}/external/libs/linux)
@@ -39,97 +56,65 @@ link_directories(
 if(${ANDROID})
   if(${QNN_DELEGATE}) # QCOM QNN delegate
     message(NOTICE  " Delegate via QCOM QNN..")
-    add_library(whisperax   SHARED
-        ${AXIE_SRC_DIR}/WhisperKit/Models/tflite_model.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Models/tflite_qnn_model.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Text/post_proc.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Audio/audio_input.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKit.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKitConfiguration.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKitPipeline.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/TranscribeTask.cpp
-	${AXIE_SRC_DIR}/WhisperKit/WhisperKitTranscriptionResult.cpp
-    )
+    SET(QNN_DELEGATE 1)
+    SET(GPU_DELEGATE 1)
   else() # generic TFLite GPU delegate
     message(NOTICE  " Delegate via Generic TFLite GPU..")
-    add_library(whisperax   SHARED
-        ${AXIE_SRC_DIR}/WhisperKit/Models/tflite_model.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Models/tflite_gpu_model.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Text/post_proc.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Audio/audio_input.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKit.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKitConfiguration.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKitPipeline.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/TranscribeTask.cpp
-	${AXIE_SRC_DIR}/WhisperKit/WhisperKitTranscriptionResult.cpp
-    )
+    SET(QNN_DELEGATE 0)
+    SET(GPU_DELEGATE 1)
 
   endif()
-
 else() # other generic arch, CPU only
   message(NOTICE  " Generic CPU only..")
-  add_library(whisperax   SHARED
-        ${AXIE_SRC_DIR}/WhisperKit/Models/tflite_model.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Text/post_proc.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/Audio/audio_input.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKit.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKitConfiguration.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/WhisperKitPipeline.cpp
-        ${AXIE_SRC_DIR}/WhisperKit/TranscribeTask.cpp
-	${AXIE_SRC_DIR}/WhisperKit/WhisperKitTranscriptionResult.cpp
-  )
+  SET(QNN_DELEGATE 0)
+  SET(GPU_DELEGATE 0)
 endif()
 
-target_include_directories(whisperax PRIVATE
-  ${AXIE_INC_DIR}
-  ${EXT_INC_DIR}
-  ${TENSORFLOW_SOURCE_DIR}
+# Delegates (QNN, GPU, ...) are required to be defined, and values are required to be 0 or 1
+if(NOT (DEFINED QNN_DELEGATE AND QNN_DELEGATE IN_LIST DELEGATE_ALLOWED_VALUES))
+    message(FATAL_ERROR "QNN_DELEGATE must be defined and set to either 0 or 1. Current value: '${QNN_DELEGATE}'")
+endif()
+
+if(NOT (DEFINED GPU_DELEGATE AND GPU_DELEGATE IN_LIST DELEGATE_ALLOWED_VALUES))
+    message(FATAL_ERROR "GPU_DELEGATE must be defined and set to either 0 or 1. Current value: '${GPU_DELEGATE}'")
+endif()
+
+add_library(whisperkit   SHARED
+    ${WHISPERKIT_SOURCES}
 )
+
+target_include_directories(whisperkit
+  PUBLIC ${WHISPERKIT_PUBLIC_HEADERS}
+  PRIVATE ${WHISPERKIT_INTERNAL_HEADERS_DIRECTORIES}
+  PRIVATE ${WHISPERKIT_EXTERNAL_HEADERS}
+  PRIVATE ${EXT_INC_DIR}
+  PRIVATE ${TENSORFLOW_SOURCE_DIR}
+)
+
+set(AUDIO_CODEC_LIBRARIES avformat avcodec avutil swresample)
+set(BASE_LIBRARIES tensorflowlite SDL3)
+set(ALL_LINKED_LIBRARIES ${BASE_LIBRARIES} ${AUDIO_CODEC_LIBRARIES})
 
 if(${ANDROID})
 
   if(${QNN_DELEGATE})
-    ADD_DEFINITIONS(-DQNN_DELEGATE)
     add_library(qnn_delegate SHARED IMPORTED)
     set_property(TARGET qnn_delegate PROPERTY IMPORTED_LOCATION
       ${EXT_LIB_DIR}/libQnnTFLiteDelegate.so
     )
-    target_link_libraries(whisperax PRIVATE
-      qnn_delegate
-      tensorflowlite
-      tensorflowlite_gpu_delegate
-      ${android_log}
-      SDL3
-      avformat
-      avcodec
-      avutil
-      swresample
-    )
+
+    list(APPEND ALL_LINKED_LIBRARIES qnn_delegate tensorflowlite_gpu_delegate ${ANDROID_LOG})
 
   else()
-    ADD_DEFINITIONS(-DGPU_DELEGATE)
-    target_link_libraries(whisperax PRIVATE
-      tensorflowlite
-      tensorflowlite_gpu_delegate
-      ${android_log}
-      SDL3
-      avformat
-      avcodec
-      avutil
-      swresample
-    )
-  endif()
 
-else()
-  target_link_libraries(whisperax PRIVATE
-    tensorflowlite
-    SDL3
-    avformat
-    avcodec
-    avutil
-    swresample
-  )
+    list(APPEND ALL_LINKED_LIBRARIES tensorflowlite_gpu_delegate ${ANDROID_LOG})
+
+  endif()
 endif()
+
+ADD_DEFINITIONS(-DGPU_DELEGATE=${GPU_DELEGATE} -DQNN_DELEGATE=${QNN_DELEGATE})
+
+target_link_libraries(whisperkit PRIVATE ${ALL_LINKED_LIBRARIES})
 
 execute_process(
   COMMAND git rev-parse HEAD
@@ -139,16 +124,16 @@ execute_process(
 )
 
 add_executable(whisperkit-cli
-    ${AXIE_CLI_DIR}/whisperkit_cli.cpp
+    ${WHISPERKIT_CLI_DIR}/whisperkit_cli.cpp
 )
 
 target_include_directories(whisperkit-cli PRIVATE
-  ${AXIE_INC_DIR}
-  ${AXIE_SRC_DIR}/external/
+  ${WHISPERKIT_PUBLIC_HEADERS}
+  ${WHISPERKIT_CLI_DIR}/external/
 )
 
 target_link_libraries(whisperkit-cli PRIVATE
-  whisperax
+  whisperkit
 )
 
 target_compile_definitions(whisperkit-cli PRIVATE
