+++ WhisperKit/src/Core/backend_class.hpp
@@ -2,8 +2,8 @@
 //  Copyright © 2024 Argmax, Inc. All rights reserved.
 #pragma once 
 
-#if defined(QNN_DELEGATE)
-#include <tflite_qnn_model.hpp>
+#if QNN_DELEGATE
+#include "tflite_qnn_model.hpp"
 #define MODEL_SUPER_CLASS     TFLiteQNN
 #else 
 
@@ -11,14 +11,14 @@
 #define kGpuBackend 1
 #define kHtpBackend 2
 
-#if defined(GPU_DELEGATE)   
+#if GPU_DELEGATE
 
-#include <tflite_gpu_model.hpp>
+#include "tflite_gpu_model.hpp"
 #define MODEL_SUPER_CLASS     TFLiteGPU
 
 #else
 
-#include <tflite_model.hpp>
+#include "tflite_model.hpp"
 #define MODEL_SUPER_CLASS     TFLiteModel
 
 #endif
