+++ WhisperKit/src/Models/tflite_gpu_model.cpp
@@ -1,6 +1,8 @@
 //  For licensing see accompanying LICENSE file.
 //  Copyright © 2024 Argmax, Inc. All rights reserved.
-#include <tflite_gpu_model.hpp>
+#if defined(GPU_DELEGATE)
+#if GPU_DELEGATE
+#include "tflite_gpu_model.hpp"
 #include <filesystem>   // C++ 17 or later
 #include "tensorflow/lite/optional_debug_tools.h"
 
@@ -75,3 +77,5 @@ bool TFLiteGPU::create_interpreter_delegate(string model_path)
 
     return true;
 }
+#endif
+#endif
