+++ WhisperKit/src/Models/tflite_qnn_model.hpp
@@ -1,6 +1,10 @@
 //  For licensing see accompanying LICENSE file.
 //  Copyright © 2024 Argmax, Inc. All rights reserved.
-#include <tflite_model.hpp>
+#pragma once
+
+#if defined(QNN_DELEGATE)
+#if QNN_DELEGATE
+#include "tflite_model.hpp"
 #include "QnnTFLiteDelegate.h"
 #include "tensorflow/lite/delegates/gpu/delegate.h"
 
@@ -23,3 +27,5 @@ protected:
 
     virtual bool create_interpreter_delegate(std::string model_path);
 };
+#endif
+#endif
