+++ WhisperKit/src/WhisperKitConfiguration.hpp
