+++ WhisperKit/src/WhisperKitTranscriptionResult.hpp
